Azurechat demo app (#57)
I've added a modified version of AzureChat that works with the
Enterprise Azure AI architecture.

Things to consider when deploying this (this is optional and AZD will
prompt for installation)
- additional resources: a Key Vault, an App Service Plan and WebApp and
a Cosmos Db account
- it re-uses the App Configuration Service that the proxy also uses
- you need permissions to register apps in Azure AD
- Azure CLI is a prerequisite for the additional scripts.
- App runs on a public endpoint, but has VNET integration to be able to
connect to the resources it needs

Note: this will delete the appregistration. The keyvault will get the
keys from the app registration only once! The script will check if the
app exists. If it exists, no changes are made (and therefor Secret
cannot be retrieved).
```
azd down --purge 
```

Let me know what you think.

original AzureChat repo
https://github.com/microsoft/azurechat

---------

Co-authored-by: Remco Brosky <remco.brosky@microsoft.com>